Move datetime helpers into separate file
