+++ sport/app/football/controllers/MatchController.scala
@@ -5,6 +5,7 @@ import feed.CompetitionsService
 import implicits.{Football, Requests}
 import model.Cached.WithoutRevalidationResult
 import model.TeamMap.findTeamIdByUrlName
+import football.datetime.DateHelpers
 import model._
 import org.joda.time.format.DateTimeFormat
 import pa.{FootballMatch, LineUp, LineUpTeam, MatchDayTeam}
